Making innerListStatus to return RemoteIterator of S3AFileStatus
